requirements.txt
Django==5.2.2
djangorestframework ==3.16.0
python-dotenv==1.1.0
psycopg2-binary==2.9.10
pytest
pytest-django
rq==1.16
redis==5.0.3
celery==5.3.6
drf-spectacular==0.28.0